Harden built-in devtools
