+++ lib/devtools/QueriesTab.tsx
@@ -1,27 +1,9 @@
-import {
-  useEffect,
-  useState,
-  type CSSProperties,
-  type KeyboardEvent as ReactKeyboardEvent,
-  type MouseEvent as ReactMouseEvent,
-  type ReactNode,
-} from 'react'
-import type { QueryRecord } from './collector.js'
-import { compactJson, formatAge, prettyJson } from './format.js'
-import type {
-  DevtoolsModel,
-  DevtoolsOperation,
-  QueryComposition,
-  QueryOwner,
-  UnderlyingFetch,
-} from './model.js'
-import {
-  Badge,
-  DetailsPane,
-  useDetailsPaneWidth,
-  useDevtoolsTheme,
-  type DevtoolsColors,
-} from './ui.js'
+import { useState, type CSSProperties } from 'react'
+import { compactJson, formatAge } from './format.js'
+import type { DevtoolsModel, DevtoolsOperation, QuerySummary } from './model.js'
+import { QueryDetails } from './QueryDetails.js'
+import { ClassBadge, plural, QueryStatusDot } from './QueryPresentation.js'
+import { useDetailsPaneWidth, useDevtoolsTheme } from './ui.js'
 
 interface QueryRow {
   operation: DevtoolsOperation
@@ -71,16 +53,6 @@ const QUERY_COLUMNS = [
   },
 ] as const
 
-const QUERY_CLASS_DESCRIPTIONS: Record<string, string> = {
-  'local-exact':
-    'Figbird can prove membership and ordering from local data, so realtime events merge directly into this result.',
-  'server-window':
-    'This is a limited or sorted server result. Figbird merges provable changes and refetches when a change could affect the window.',
-  'server-authoritative':
-    'The server decides which records belong in this result. Realtime changes trigger a server refetch.',
-  get: 'A direct record lookup by ID.',
-}
-
 export function QueriesTab({
   model,
   filter,
@@ -96,11 +68,8 @@ export function QueriesTab({
   const [selectedOperationKey, setSelectedOperationKey] = useState<string | null>(null)
   const [detailsWidth, onDetailsResizeStart] = useDetailsPaneWidth()
   const rows: QueryRow[] = model.operations.flatMap(operation => {
-    const query = operation.query
-    const localSubscriberCount = Array.from(operation.relationalKeys).reduce(
-      (count, key) => count + (inspectedQueryCounts?.get(key) ?? 0),
-      0,
-    )
+    const query = operation.summary
+    const localSubscriberCount = inspectedQueryCounts?.get(operation.key) ?? 0
     if (inspectedQueryCounts && localSubscriberCount === 0) return []
     if (
       activeOnly &&
@@ -116,7 +85,7 @@ export function QueriesTab({
       JSON.stringify(query.query ?? {}),
       ...operation.underlying.map(item =>
         [
-          item.owner.path,
+          item.path,
           item.query.serviceName,
           item.query.method,
           JSON.stringify(item.query.query ?? {}),
@@ -134,7 +103,7 @@ export function QueriesTab({
     textOverflow: 'ellipsis',
     whiteSpace: 'nowrap',
   }
-  const selectedOperation = rows.find(row => row.operation.key === selectedOperationKey)
+  const selectedOperation = rows.find(row => row.operation.key === selectedOperationKey)?.operation
 
   return (
     <section style={{ height: '100%', display: 'flex', minWidth: 0 }}>
@@ -169,9 +138,8 @@ export function QueriesTab({
                 </td>
               </tr>
             ) : null}
-            {rows.map(row => {
-              const { operation } = row
-              const query = operation.query
+            {rows.map(({ operation, localSubscriberCount }) => {
+              const query = operation.summary
               const isSelected = selectedOperationKey === operation.key
               return (
                 <tr
@@ -231,10 +199,10 @@ export function QueriesTab({
                       </span>
                       {inspectedQueryCounts ? (
                         <span
-                          title={`${row.localSubscriberCount} of ${query.subscriberCount} subscribers in the selected area`}
+                          title={`${localSubscriberCount} of ${query.subscriberCount} subscribers in the selected area`}
                           style={{ color: colors.blue, whiteSpace: 'nowrap', flexShrink: 0 }}
                         >
-                          {row.localSubscriberCount} here
+                          {localSubscriberCount} here
                         </span>
                       ) : null}
                     </span>
@@ -284,531 +252,27 @@ export function QueriesTab({
       </div>
       {selectedOperation ? (
         <QueryDetails
-          query={selectedOperation.operation.query}
-          rootFetches={selectedOperation.operation.rootFetches}
-          underlying={selectedOperation.operation.underlying}
+          operation={selectedOperation}
           width={detailsWidth}
           onResizeStart={onDetailsResizeStart}
           onClose={() => setSelectedOperationKey(null)}
-          {...(selectedOperation.operation.composition
-            ? { composition: selectedOperation.operation.composition }
-            : {})}
         />
       ) : null}
     </section>
   )
 }
 
-function QueryDetails({
-  query,
-  rootFetches,
-  underlying,
-  composition,
-  width,
-  onResizeStart,
-  onClose,
-}: {
-  query: QueryRecord
-  rootFetches: QueryRecord[]
-  underlying: UnderlyingFetch[]
-  composition?: QueryComposition
-  width: number
-  onResizeStart: (event: ReactMouseEvent<HTMLDivElement>) => void
-  onClose: () => void
-}) {
-  const { colors, styles } = useDevtoolsTheme()
-  const [selectedUnderlyingKey, setSelectedUnderlyingKey] = useState<string | null>(null)
-  useEffect(() => setSelectedUnderlyingKey(null), [query.queryId])
-  const selectedUnderlying =
-    underlying.find(item => underlyingFetchKey(item) === selectedUnderlyingKey) ?? null
-  const activeQuery = selectedUnderlying?.query ?? query
-  const average =
-    activeQuery.fetchCount > 0 ? activeQuery.totalDurationMs / activeQuery.fetchCount : 0
-  const status = queryStatus(activeQuery)
-  const rootTitle = `${query.serviceName}.${query.method}`
-  const rootOperation = composition?.operation ?? rootTitle
-  const activeOperation = `${activeQuery.serviceName}.${activeQuery.method}`
-  const rootQueryIdentity =
-    rootFetches.length === 1 ? rootFetches[0]!.queryId : `${rootFetches.length} root query IDs`
-  const directChildren = selectedUnderlying
-    ? underlying.filter(item => {
-        if (item.owner.label !== selectedUnderlying.owner.label) return false
-        const prefix = `${selectedUnderlying.owner.path}.`
-        if (!item.owner.path.startsWith(prefix)) return false
-        return !item.owner.path.slice(prefix.length).includes('.')
-      })
-    : underlying
-  const breadcrumb = selectedUnderlying ? (
-    <>
-      <button
-        type='button'
-        aria-label='Back to root query'
-        onClick={() => setSelectedUnderlyingKey(null)}
-        style={breadcrumbButtonStyle(colors)}
-      >
-        {rootTitle}
-      </button>
-      {selectedUnderlying.owner.path.split('.').map((segment, index, segments) => {
-        const path = segments.slice(0, index + 1).join('.')
-        const ancestor = underlying.find(
-          item => item.owner.label === selectedUnderlying.owner.label && item.owner.path === path,
-        )
-        const current = index === segments.length - 1
-        return (
-          <span key={`${selectedUnderlying.owner.label}:${path}`} style={{ display: 'contents' }}>
-            <span style={{ color: colors.faint }}>›</span>
-            {current || !ancestor ? (
-              <span>{segment}</span>
-            ) : (
-              <button
-                type='button'
-                onClick={() => setSelectedUnderlyingKey(underlyingFetchKey(ancestor))}
-                style={breadcrumbButtonStyle(colors)}
-              >
-                {segment}
-              </button>
-            )}
-          </span>
-        )
-      })}
-    </>
-  ) : (
-    rootTitle
-  )
-  return (
-    <DetailsPane
-      title={breadcrumb}
-      subtitle={
-        <code
-          title={[
-            'Cache identity derived from the operation and parameters. Changing filters creates a new query ID.',
-            rootFetches.length > 1 ? rootFetches.map(root => root.queryId).join('\n') : undefined,
-          ]
-            .filter(Boolean)
-            .join('\n\n')}
-          style={{
-            ...styles.code,
-            color: colors.muted,
-            overflow: 'hidden',
-            textOverflow: 'ellipsis',
-            whiteSpace: 'nowrap',
-          }}
-        >
-          {selectedUnderlying ? `${activeOperation} · ${activeQuery.queryId}` : rootQueryIdentity}
-        </code>
-      }
-      width={width}
-      onResizeStart={onResizeStart}
-      onClose={onClose}
-    >
-      <div
-        style={{
-          display: 'flex',
-          alignItems: 'center',
-          gap: 7,
-          minHeight: 24,
-          marginBottom: 10,
-        }}
-      >
-        <QueryStatusDot query={activeQuery} />
-        <span style={{ color: colors.text, fontWeight: 600 }}>{status.kind}</span>
-        <ClassBadge value={activeQuery.classification} />
-        <span style={styles.spacer} />
-        <span style={{ color: colors.muted, whiteSpace: 'nowrap' }}>
-          {plural(activeQuery.subscriberCount, 'subscriber', 'subscribers')}
-        </span>
-      </div>
-      {activeQuery.lastError ? (
-        <div
-          style={{
-            color: colors.red,
-            background: colors.panel2,
-            borderLeft: `3px solid ${colors.red}`,
-            padding: '7px 9px',
-            marginBottom: 10,
-          }}
-        >
-          {activeQuery.lastError.message}
-        </div>
-      ) : null}
-      <div
-        style={{
-          display: 'grid',
-          gridTemplateColumns: 'repeat(2, minmax(0, 1fr))',
-          borderTop: `1px solid ${colors.rowBorder}`,
-          borderBottom: `1px solid ${colors.rowBorder}`,
-          marginBottom: 14,
-        }}
-      >
-        <QueryMetric value={String(activeQuery.itemCount)} label='rows' />
-        <QueryMetric value={String(activeQuery.fetchCount)} label='fetches' borderLeft />
-        <QueryMetric
-          value={
-            activeQuery.lastDurationMs === undefined
-              ? '-'
-              : `${Math.round(activeQuery.lastDurationMs)}ms`
-          }
-          label='last fetch'
-          borderTop
-        />
-        <QueryMetric
-          value={activeQuery.fetchCount === 0 ? '-' : `${Math.round(average)}ms`}
-          label='average'
-          borderLeft
-          borderTop
-        />
-      </div>
-      {activeQuery.spans.length > 0 ? (
-        <QueryDetailSection
-          label='Recent fetches'
-          meta={`${Math.round(activeQuery.totalDurationMs)}ms total`}
-        >
-          <Sparkline spans={activeQuery.spans} />
-        </QueryDetailSection>
-      ) : null}
-      {!selectedUnderlying || directChildren.length > 0 ? (
-        <QueryDetailSection
-          label={selectedUnderlying ? 'Related queries' : 'Query plan'}
-          {...(directChildren.length > 0
-            ? { meta: plural(directChildren.length, 'relation', 'relations') }
-            : {})}
-        >
-          {!selectedUnderlying ? (
-            <QueryPlanRow
-              path='root'
-              operation={rootOperation}
-              detail={
-                composition?.planDetail ??
-                (query.method === 'get'
-                  ? `id ${query.resourceId ?? '?'}`
-                  : compactJson(query.query ?? {}))
-              }
-            />
-          ) : null}
-          {directChildren.length > 0 ? (
-            <div
-              style={{
-                marginLeft: selectedUnderlying ? 0 : 7,
-                paddingLeft: selectedUnderlying ? 0 : 13,
-                borderLeft: selectedUnderlying ? undefined : `1px solid ${colors.border}`,
-              }}
-            >
-              {directChildren.map(item => (
-                <QueryPlanRow
-                  key={underlyingFetchKey(item)}
-                  path={
-                    selectedUnderlying
-                      ? (item.owner.path.split('.').pop() ?? item.owner.path)
-                      : formatOwnerPath(item.owner)
-                  }
-                  operation={`${item.query.serviceName}.${item.query.method}`}
-                  detail={[
-                    plural(item.query.itemCount, 'row', 'rows'),
-                    item.query.lastDurationMs === undefined
-                      ? undefined
-                      : `${Math.round(item.query.lastDurationMs)}ms`,
-                  ]
-                    .filter(Boolean)
-                    .join(' · ')}
-                  classification={item.query.classification}
-                  onSelect={() => setSelectedUnderlyingKey(underlyingFetchKey(item))}
-                />
-              ))}
-            </div>
-          ) : null}
-        </QueryDetailSection>
-      ) : null}
-      <details
-        key={activeQuery.queryId}
-        style={{
-          borderTop: `1px solid ${colors.rowBorder}`,
-          padding: '11px 0',
-        }}
-      >
-        <summary
-          style={{
-            color: colors.text,
-            cursor: 'pointer',
-            fontWeight: 600,
-          }}
-        >
-          Parameters
-          <code
-            title={prettyJson(activeQuery.query ?? {})}
-            style={{
-              ...styles.code,
-              display: 'inline-block',
-              maxWidth: 'calc(100% - 90px)',
-              color: colors.muted,
-              fontWeight: 400,
-              marginLeft: 8,
-              overflow: 'hidden',
-              textOverflow: 'ellipsis',
-              verticalAlign: 'bottom',
-              whiteSpace: 'nowrap',
-            }}
-          >
-            {compactJson(activeQuery.query ?? {})}
-          </code>
-        </summary>
-        <pre
-          style={{
-            ...styles.code,
-            whiteSpace: 'pre-wrap',
-            overflowWrap: 'anywhere',
-            margin: '9px 0 0',
-            padding: 10,
-            background: colors.panel2,
-            borderRadius: 4,
-          }}
-        >
-          {prettyJson(activeQuery.query ?? {})}
-        </pre>
-      </details>
-    </DetailsPane>
-  )
-}
-
-function underlyingFetchKey(item: UnderlyingFetch): string {
-  return `${item.owner.label}:${item.owner.path}:${item.query.queryId}`
-}
-
-function breadcrumbButtonStyle(colors: DevtoolsColors): CSSProperties {
-  return {
-    border: 0,
-    background: 'transparent',
-    color: colors.blue,
-    padding: 0,
-    font: 'inherit',
-    fontWeight: 'inherit',
-    cursor: 'pointer',
-  }
-}
-
-function QueryMetric({
-  value,
-  label,
-  borderLeft,
-  borderTop,
-}: {
-  value: string
-  label: string
-  borderLeft?: boolean
-  borderTop?: boolean
-}) {
-  const { colors } = useDevtoolsTheme()
-  return (
-    <div
-      style={{
-        display: 'flex',
-        alignItems: 'baseline',
-        gap: 5,
-        minWidth: 0,
-        padding: '9px 0',
-        paddingLeft: borderLeft ? 12 : 0,
-        borderLeft: borderLeft ? `1px solid ${colors.rowBorder}` : undefined,
-        borderTop: borderTop ? `1px solid ${colors.rowBorder}` : undefined,
-      }}
-    >
-      <strong style={{ color: colors.text, fontSize: 14 }}>{value}</strong>
-      <span style={{ color: colors.muted, whiteSpace: 'nowrap' }}>{label}</span>
-    </div>
-  )
-}
-
-function QueryDetailSection({
-  label,
-  meta,
-  children,
-}: {
-  label: string
-  meta?: string
-  children: ReactNode
-}) {
-  const { colors } = useDevtoolsTheme()
-  return (
-    <section style={{ marginBottom: 14 }}>
-      <div
-        style={{
-          display: 'flex',
-          alignItems: 'baseline',
-          gap: 8,
-          marginBottom: 5,
-        }}
-      >
-        <strong style={{ color: colors.text, fontWeight: 650 }}>{label}</strong>
-        {meta ? <span style={{ color: colors.muted }}>{meta}</span> : null}
-      </div>
-      {children}
-    </section>
-  )
-}
-
-function QueryPlanRow({
-  path,
-  operation,
-  detail,
-  classification,
-  onSelect,
-}: {
-  path: string
-  operation: string
-  detail: string
-  classification?: string
-  onSelect?: () => void
-}) {
-  const { colors, styles } = useDevtoolsTheme()
-  return (
-    <div
-      {...(onSelect
-        ? {
-            role: 'button',
-            tabIndex: 0,
-            'aria-label': `Inspect nested query ${path}`,
-            title: `Inspect ${path}`,
-            onClick: onSelect,
-            onKeyDown: (event: ReactKeyboardEvent<HTMLDivElement>) => {
-              if (event.key === 'Enter' || event.key === ' ') {
-                event.preventDefault()
-                onSelect()
-              }
-            },
-          }
-        : {})}
-      style={{
-        display: 'grid',
-        gridTemplateColumns: 'minmax(0, 1fr) auto',
-        alignItems: 'start',
-        gap: 8,
-        padding: '7px 0',
-        borderTop: `1px solid ${colors.rowBorder}`,
-        cursor: onSelect ? 'pointer' : undefined,
-      }}
-    >
-      <div style={{ minWidth: 0 }}>
-        <div
-          style={{
-            display: 'flex',
-            alignItems: 'baseline',
-            gap: 7,
-            minWidth: 0,
-          }}
-        >
-          <code style={{ ...styles.code, color: colors.faint, flexShrink: 0 }}>{path}</code>
-          <strong
-            style={{
-              overflow: 'hidden',
-              textOverflow: 'ellipsis',
-              whiteSpace: 'nowrap',
-              fontWeight: 600,
-            }}
-          >
-            {operation}
-          </strong>
-        </div>
-        <div
-          title={detail}
-          style={{
-            ...styles.code,
-            color: colors.muted,
-            overflow: 'hidden',
-            textOverflow: 'ellipsis',
-            whiteSpace: 'nowrap',
-            marginTop: 3,
-          }}
-        >
-          {detail || 'all'}
-        </div>
-      </div>
-      {classification || onSelect ? (
-        <div style={{ display: 'flex', alignItems: 'center', gap: 7 }}>
-          {classification ? <ClassBadge value={classification} /> : null}
-          {onSelect ? (
-            <span
-              aria-hidden='true'
-              style={{ color: colors.blue, fontSize: 16, lineHeight: '18px' }}
-            >
-              ›
-            </span>
-          ) : null}
-        </div>
-      ) : null}
-    </div>
-  )
-}
-
-function QueryStatusDot({ query }: { query: QueryRecord }) {
-  const { colors } = useDevtoolsTheme()
-  const status = queryStatus(query)
-  const color =
-    status.kind === 'error'
-      ? colors.red
-      : status.kind === 'fetching'
-        ? colors.blue
-        : status.kind === 'cached'
-          ? colors.amber
-          : status.kind === 'active'
-            ? colors.green
-            : colors.muted
-  return (
-    <span
-      title={status.label}
-      aria-label={status.label}
-      style={{
-        width: 8,
-        height: 8,
-        borderRadius: 999,
-        background: color,
-        boxShadow: `0 0 0 2px ${colors.panel}`,
-        flexShrink: 0,
-      }}
-    />
-  )
-}
-
-function queryStatus(query: QueryRecord): {
-  kind: 'active' | 'cached' | 'error' | 'fetching' | 'idle'
-  label: string
-} {
-  if (query.status === 'error') {
-    return { kind: 'error', label: statusLabel(query, 'error') }
-  }
-
-  if (query.isFetching || query.status === 'loading') {
-    return { kind: 'fetching', label: statusLabel(query, 'fetching') }
-  }
-
-  if (query.subscriberCount === 0) {
-    return { kind: 'cached', label: statusLabel(query, 'cached') }
-  }
-
-  if (query.subscriberCount > 0) return { kind: 'active', label: statusLabel(query, 'active') }
-  return { kind: 'idle', label: statusLabel(query, 'idle') }
-}
-
-function statusLabel(query: QueryRecord, state: string): string {
-  const subscribers =
-    query.subscriberCount === 1 ? '1 subscriber' : `${query.subscriberCount} subscribers`
-  return `${state} · ${subscribers} · ${query.status}`
-}
-
-function QueryRows({ query }: { query: QueryRecord }) {
+function QueryRows({ query }: { query: QuerySummary }) {
   const { colors } = useDevtoolsTheme()
   return <span style={{ color: colors.text, fontWeight: 600 }}>{query.itemCount}</span>
 }
 
-function QueryFetchStats({ query }: { query: QueryRecord }) {
+function QueryFetchStats({ query }: { query: QuerySummary }) {
   const { colors } = useDevtoolsTheme()
   const parts = [plural(query.fetchCount, 'fetch', 'fetches')]
-  if (query.realtimeSeen > 0) {
-    parts.push(plural(query.realtimeSeen, 'event', 'events'))
-  }
-  if (query.reconciles > 0) {
-    parts.push(plural(query.reconciles, 'refetch', 'refetches'))
-  }
-  if (query.errorCount > 0) {
-    parts.push(plural(query.errorCount, 'error', 'errors'))
-  }
+  if (query.realtimeSeen > 0) parts.push(plural(query.realtimeSeen, 'event', 'events'))
+  if (query.reconciles > 0) parts.push(plural(query.reconciles, 'refetch', 'refetches'))
+  if (query.errorCount > 0) parts.push(plural(query.errorCount, 'error', 'errors'))
 
   return (
     <span
@@ -826,11 +290,7 @@ function QueryFetchStats({ query }: { query: QueryRecord }) {
   )
 }
 
-function plural(count: number, singular: string, pluralValue: string): string {
-  return `${count} ${count === 1 ? singular : pluralValue}`
-}
-
-function QueryLastTiming({ query }: { query: QueryRecord }) {
+function QueryLastTiming({ query }: { query: QuerySummary }) {
   const { colors } = useDevtoolsTheme()
   const duration =
     query.lastDurationMs === undefined
@@ -861,7 +321,7 @@ function QueryLastTiming({ query }: { query: QueryRecord }) {
   )
 }
 
-function QueryAge({ query }: { query: QueryRecord }) {
+function QueryAge({ query }: { query: QuerySummary }) {
   const { colors } = useDevtoolsTheme()
   return (
     <span
@@ -877,49 +337,4 @@ function QueryAge({ query }: { query: QueryRecord }) {
       {query.fetchedAt ? formatAge(Date.now() - query.fetchedAt) : '-'}
     </span>
   )
-}
-
-function ClassBadge({ value }: { value: string }) {
-  const tone =
-    value === 'local-exact'
-      ? 'green'
-      : value === 'server-window'
-        ? 'amber'
-        : value === 'get'
-          ? 'neutral'
-          : 'red'
-  return (
-    <Badge tone={tone} title={QUERY_CLASS_DESCRIPTIONS[value]}>
-      {value}
-    </Badge>
-  )
-}
-
-function Sparkline({ spans }: { spans: QueryRecord['spans'] }) {
-  const { colors } = useDevtoolsTheme()
-  if (spans.length === 0) return null
-  const max = Math.max(1, ...spans.map(span => (span.endAt ?? span.startAt) - span.startAt))
-  return (
-    <div style={{ display: 'flex', alignItems: 'end', gap: 2, height: 28, marginTop: 8 }}>
-      {spans.slice(-30).map((span, index) => {
-        const duration = (span.endAt ?? span.startAt) - span.startAt
-        return (
-          <span
-            key={`${span.startAt}:${index}`}
-            title={`${Math.round(duration)}ms`}
-            style={{
-              width: 4,
-              height: Math.max(3, (duration / max) * 24),
-              background: span.ok === false ? colors.red : colors.green,
-              borderRadius: 2,
-            }}
-          />
-        )
-      })}
-    </div>
-  )
-}
-
-function formatOwnerPath(owner: QueryOwner): string {
-  return owner.path === '(root)' ? 'root' : owner.path
 }
